index + worldmap feature update
Added sign-in button on splash page.
Setup function to check if user is signed in and if so a welcome screen
appears to welcome them back and remove back button to reduce loop
error.
Added Temporary modal for user icon in /worldmap.tsx to allow user to
signout and return to splash page /index.tsx
TODO: add startpack page in onboarding and resolve account.tsx buttons