Refactor Attempt - Update
Refactoring is complete. Now adding JSDoc for functions, Hooks and
components for everyone to understand the code and sequence and flow. We
dont have to apply this to the code base. however I feel having it will
teach us how JSDocs work and how we can communicate through our code.
TODO: createPin.tsx is comeplete. Need to update the remaining Hooks and
components.